+++ src/components/Dialog/index.js
@@ -1,32 +1,35 @@
 import PropTypes from 'prop-types';
 import React from 'react';
-import Icon from 'react-oui-icons';
 import ButtonRow from '../ButtonRow';
+import CloseButton from '../CloseButton';
+import Fieldset from '../Fieldset';
+import Footer from './Deprecated/Footer';
+import Title from './Deprecated/Title';
+import Wrapper from './Deprecated/Wrapper';
 
-const Dialog = props => (
+export const DialogNew = props => (
   <div className="oui-dialog__wrapper">
     <div className="oui-dialog__overlay" />
     <div
       data-ui-component={ true }
-      className="oui-dialog overflow-y--auto"
+      className="oui-dialog"
       data-test-section={ props.testSection }>
-      {props.hasCloseButton && <button className="oui-dialog__close" onClick={ props.onClose }>
-        <Icon name="close" />
-      </button>}
-      <div className="oui-dialog__header">
-        <h2 className="push--bottom">{props.title}</h2>
-      </div>
-      <div className="oui-dialog__body">
+      {props.hasCloseButton && <CloseButton onClick={ props.onClose } size="medium"/>}
+      <header className="oui-dialog__header">
+        <h2 className="flush--bottom">{props.title}</h2>
+        {props.subtitle && <p className="push--top flush--bottom">{props.subtitle}</p>}
+      </header>
+      <div className="oui-dialog__body overflow-y--auto">
         {props.children}
       </div>
-      <div className="oui-dialog__footer">
-        <ButtonRow rightGroup={ props.footerButtonContent }/>
-      </div>
+      <footer className="oui-dialog__footer-container">
+        <ButtonRow rightGroup={ props.footerButtonList }/>
+      </footer>
     </div>
   </div>
 );
 
-Dialog.propTypes = {
+DialogNew.propTypes = {
   /**
    * The body of the dialog to request minimal information from the user.
    */
@@ -34,7 +37,7 @@ Dialog.propTypes = {
   /**
    * Array of buttons used in the footer of the dialog.
    */
-  footerButtonContent: PropTypes.array.isRequired,
+  footerButtonList: PropTypes.array.isRequired,
   /**
    *  Used to determine if the dialog should have a close button.
    */
@@ -44,6 +47,10 @@ Dialog.propTypes = {
    */
   onClose: PropTypes.func,
   /**
+   * A subtitle for the dialog.
+   */
+  subtitle: PropTypes.string,
+  /**
    * Identifier used to create data-test-section attributes for testing.
    */
   testSection: PropTypes.string,
@@ -51,6 +58,20 @@ Dialog.propTypes = {
    * Main title of the dialog.
    */
   title: PropTypes.string.isRequired,
+};
+
+DialogNew.defaultProps = {
+  hasCloseButton: true,
+  onClose: () => {},
+  subtitle: '',
+  testSection: '',
+};
+
+const Dialog = {
+  Fieldset,
+  Footer,
+  Title,
+  Wrapper,
 };
 
 export default Dialog;